+++ packages/cryptoplease/lib/features/incoming_split_key_payment/bl/bloc.dart
@@ -1,6 +1,6 @@
 import 'package:bloc_concurrency/bloc_concurrency.dart';
-import 'package:cryptoplease/core/api_version.dart';
 import 'package:cryptoplease/core/processing_state.dart';
+import 'package:cryptoplease/core/split_key_api_version.dart';
 import 'package:cryptoplease/data/transaction/tx_creator.dart';
 import 'package:cryptoplease/data/transaction/tx_creator_selector.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/models.dart';
