+++ packages/cryptoplease/lib/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/ft/bloc.dart
@@ -1,8 +1,8 @@
 import 'package:cryptoplease/core/amount.dart';
-import 'package:cryptoplease/core/api_version.dart';
 import 'package:cryptoplease/core/conversion_rates/bl/repository.dart';
 import 'package:cryptoplease/core/currency.dart';
 import 'package:cryptoplease/core/flow.dart';
+import 'package:cryptoplease/core/split_key_api_version.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/repository.dart';
