+++ packages/cryptoplease/lib/features/outgoing_transfer/bl/outgoing_payment.dart
@@ -1,8 +1,8 @@
 import 'package:collection/collection.dart';
 import 'package:cryptoplease/config.dart';
 import 'package:cryptoplease/core/amount.dart';
-import 'package:cryptoplease/core/api_version.dart';
 import 'package:cryptoplease/core/currency.dart';
+import 'package:cryptoplease/core/split_key_api_version.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/tokens/token_list.dart';
 import 'package:cryptoplease/core/wallet.dart';
