+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/fungible_token/link_transfer_flow_screen.dart
@@ -1,9 +1,9 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/app/routes.dart';
 import 'package:cryptoplease/core/amount.dart';
-import 'package:cryptoplease/core/api_version.dart';
 import 'package:cryptoplease/core/balances/bl/balances_bloc.dart';
 import 'package:cryptoplease/core/conversion_rates/bl/repository.dart';
+import 'package:cryptoplease/core/split_key_api_version.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/user_preferences.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/ft/bloc.dart';
